pyproject.toml
[project]
name = "fastapi-products-api"
version = "0.1.0"
description = ""
authors = [
    {name = "owhenrique",email = "me.pauloalmeida@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "fastapi[standard] (>=0.115.11,<0.116.0)",
    "sqlalchemy (>=2.0.39,<3.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "alembic (>=1.15.1,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "pwdlib[argon2] (>=0.2.1,<0.3.0)",
    "tzdata (>=2025.1,<2026.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
taskipy = "^1.14.1"
pytest = "^8.3.5"
pytest-cov = "^6.0.0"
ruff = "^0.11.0"

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT', 'FAST']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.taskipy.tasks]
lint = 'ruff check'
pre_format = 'ruff check --fix'
format = 'ruff format'
run = 'fastapi dev fastapi_products_api/app.py'
pre_test = 'task lint'
test = 'pytest -s -x --cov=fastapi_products_api -vv'
post_test = 'coverage html'